Seperate shim manager and task service
Create new shim manager interface and deprecate older
shim manager interface.
